refactor: extract plugin service
